Move more code into Common
