+++ SoundSwitch.Common/Resources/defaultMicrophone.ico
