+++ SoundSwitch.Common/Resources/defaultSpeakers.ico
