+++ SoundSwitch.Common/packages.config
@@ -1,4 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="NAudio" version="1.9.0" targetFramework="net472" />
+  <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net472" />
   <package id="Serilog" version="2.9.0" targetFramework="net472" />
 </packages>
