+++ SoundSwitch/Framework/Audio/Lister/CachedAudioDeviceLister.cs
@@ -19,7 +19,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using NAudio.CoreAudioApi;
 using Serilog;
-using SoundSwitch.Framework.Audio.Device;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Framework.NotificationManager;
 using SoundSwitch.Model;
 using SoundSwitch.Util.Timer;
