+++ SoundSwitch/Framework/Configuration/ISoundSwitchConfiguration.cs
@@ -14,8 +14,8 @@
 
 using System;
 using System.Collections.Generic;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Common.WinApi.Keyboard;
-using SoundSwitch.Framework.Audio.Device;
 using SoundSwitch.Framework.DeviceCyclerManager;
 using SoundSwitch.Framework.NotificationManager;
 using SoundSwitch.Framework.Profile;
