+++ SoundSwitch/Framework/Configuration/SoundSwitchConfiguration.cs
@@ -19,8 +19,8 @@ using System.Linq;
 using System.Windows.Forms;
 using NAudio.CoreAudioApi;
 using Serilog;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Common.WinApi.Keyboard;
-using SoundSwitch.Framework.Audio.Device;
 using SoundSwitch.Framework.DeviceCyclerManager;
 using SoundSwitch.Framework.NotificationManager;
 using SoundSwitch.Framework.Profile;
