+++ SoundSwitch/Framework/DeviceCyclerManager/DeviceCycler/DeviceCyclerAll.cs
@@ -16,7 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using NAudio.CoreAudioApi;
-using SoundSwitch.Framework.Audio.Device;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Model;
 using SoundSwitch.Localization;
