+++ SoundSwitch/Framework/DeviceCyclerManager/DeviceCycler/DeviceCyclerAvailable.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Collections.Generic;
 using NAudio.CoreAudioApi;
-using SoundSwitch.Framework.Audio.Device;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Model;
 using SoundSwitch.Localization;
