+++ SoundSwitch/Framework/Profile/ProfileManager.cs
@@ -7,7 +7,6 @@ using SoundSwitch.Audio.Manager;
 using SoundSwitch.Audio.Manager.Interop.Enum;
 using SoundSwitch.Common.WinApi;
 using SoundSwitch.Common.WinApi.Keyboard;
-using SoundSwitch.Framework.Audio.Device;
 using SoundSwitch.Framework.Configuration;
 using SoundSwitch.Localization;
