+++ SoundSwitch/Framework/TrayIcon/Icon/Changer/NeverIconIconChanger.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using SoundSwitch.Framework.Audio.Device;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Localization;
 using SoundSwitch.Properties;
