+++ SoundSwitch/Framework/TrayIcon/Icon/Changer/PlaybackIconChanger.cs
@@ -1,5 +1,5 @@
 ﻿using NAudio.CoreAudioApi;
-using SoundSwitch.Framework.Audio.Device;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Localization;
 using SoundSwitch.Model;
