+++ SoundSwitch/Framework/TrayIcon/Icon/IIconChanger.cs
@@ -1,4 +1,4 @@
-﻿using SoundSwitch.Framework.Audio.Device;
+﻿using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Framework.Factory;
 
 namespace SoundSwitch.Framework.TrayIcon.Icon
