+++ SoundSwitch/Model/Events.cs
@@ -15,8 +15,8 @@
 using System;
 using System.Collections.Generic;
 using NAudio.CoreAudioApi;
+using SoundSwitch.Common.Framework.Audio.Device;
 using SoundSwitch.Framework.Audio;
-using SoundSwitch.Framework.Audio.Device;
 using SoundSwitch.Framework.NotificationManager;
 using SoundSwitch.Framework.Updater;
